requirements.txt
aiohttp==3.6.3
appdirs==1.4.4
astroid==2.4.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
black==20.8b1
braceexpand==0.1.6
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
discord-pretty-help==1.2.1
discord.py==1.5.0
distest==0.6.0
dnspython==2.1.0
easydict==1.9
fastapi==0.63.0
Flask==1.1.2
gevent==20.9.0
greenlet==0.4.17
h11==0.12.0
humanize==3.0.1
idna==2.10
import-expression==1.1.4
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
jishaku
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.8.0
marshmallow-sqlalchemy==0.24.2
mccabe==0.6.1
MongoSanitizer==0.0.1
motor
multidict==4.7.6
mypy-extensions==0.4.3
pathspec==0.8.1
prettytable==2.0.0
pycparser==2.20
pydantic==1.7.3
pylint==2.6.0
pymongo
python-dateutil==2.8.1
python-dotenv
regex==2020.11.13
requirements-detector==0.7
six==1.15.0
SQLAlchemy==1.3.23
starlette==0.13.6
toml==0.10.2
tornado==6.0.4
typed-ast==1.4.2
typing-extensions==3.7.4.3
umongo
uvicorn==0.13.4
vermin==0.10.5
wcwidth==0.2.5
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.5.1
zope.event==4.5.0
zope.interface==5.1.2

disputils
